requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-unfold>=0.9.0
python-decouple>=3.8
psycopg2-binary>=2.9.0
Pillow>=10.0.0

